+++ src/controllers/challengeController.js
@@ -1,5 +1,6 @@
-// challengeController.js
-
+import cors from 'cors';
+import express from 'express';
+import { supabase } from '../supabaseClient.js';
 // Mock data (replace with database)
 const challenges = [
     [
@@ -45,9 +46,27 @@ const challenges = [
         { id: 40, description: 'Set intentions for the day; let each goal be a step toward a fulfilling journey.' }
       ]
     ]
-      
+    const app = express();
+    const port = 3020;
+    app.use(cors());  
+  
+    async function getData() {
+      const { data, error } = await supabase.from('challenges').select('*');
+      if (error) console.log('query error', error);
+      else{
+          console.log(data);
+          return data;
+      } 
+  }
+  export const getChallenges = async function getChallengeData(req, res, next) {
+    try {
+      res.json(await getData());
+    } catch (error) {
+      next(error);
+    }
+  }
 
-    export const getAllChallenges = async (req, res) => {
+  export const getAllChallenges = async (req, res) => {
         try {
           // Assuming challenges is some asynchronous data retrieval operation
           res.json(challenges);
